Fix warning in MeshManipulation.cpp: change variable to size_t.

+++ src/graphics/data/MeshManipulation.cpp
@@ -150,7 +150,7 @@ long IsInSelection(EERIE_3DOBJ * obj, long vert, long tw)
 
 	if (vert < 0) return -1;
 
-	for (long i = 0; i < obj->selections[tw].selected.size(); i++)
+	for (size_t i = 0; i < obj->selections[tw].selected.size(); i++)
 	{
 		if (obj->selections[tw].selected[i] == vert)
 			return i;
